Time logging/ excuses
This latest update allows teachers/ parents to view times that students
log in and out of school. An adjustment that was made to this was only
presenting records that are out of the ordinary ie. student was late,
student left early etc. This was done in an attempt to reduce the amount
of redundant content shown to a user at any particular time. Parents now
also have the option to submit excuses for these out of the ordinary
records.